Manage installation of collectd on Red Hat/Centos distributions.

+++ recipes/config.rb
@@ -7,34 +7,44 @@
 
 
 # Get collectd configuration (main params, plugins, chains).
-Chef::Log.info("get configuration from attributes")
-files = get_collectd_conf(convert(node.to_hash))
+Chef::Log.info('get configuration from attributes')
+pkgs, files = get_collectd_conf(convert(node.to_hash))
 
 # Remove everything in the directory except our configuration files.
-ruby_block 'remove_unused_files' do
-  block do
-    cur_files = Dir.entries('/etc/collectd')
-                   .select{ |path| path != '.' && path != '..' }
-                   .map{ |path| "/etc/collectd/#{path}" }
-    cur_files.each do |filepath|
-      Chef::Log.info(filepath)
-      if !files.include?(filepath)
-        Chef::Log.info("removing '#{filepath}'")
-        FileUtils.remove_entry(filepath)
+if node['collectd']['deb']['supported_platforms'].include?(node['platform'])
+  ruby_block 'remove_unused_files' do
+    block do
+      cur_files = Dir.entries('/etc/collectd')
+                     .select{ |path| path != '.' && path != '..' }
+                     .map{ |path| "/etc/collectd/#{path}" }
+      cur_files.each do |filepath|
+        if !files.include?(filepath)
+          Chef::Log.info("removing '#{filepath}'")
+          FileUtils.remove_entry(filepath)
+        end
       end
     end
+    action :run
   end
-  action :run
 end
 
-# Deploy configuration file and restart service if there is any change.
+# Install dependencies, deploy configuration file and restart
+# service if there is any change.
 service('collectd') { action :nothing }
+
+pkgs.each do |pkg|
+  package pkg do
+    action :install
+    notifies :restart, 'service[collectd]', :delayed
+  end
+end
+
 files.each do |filepath, content|
   file filepath do
     mode '0644'
     owner 'root'
     group 'root'
     content content
-    notifies :restart, "service[collectd]", :delayed
+    notifies :restart, 'service[collectd]', :delayed
   end
 end
